Make package build on any host
The makefile provided by the source code defines the targeted kernel
release by running `uname -r` but it can be overriden by setting
BUILD_KERNEL environment variable.

Also, the install target of the makefile provided with the sources
performs useful stuff to really install the module on the nodes, but
nothing needed in the package though. The dh_auto_install is overriden
to disable the install target of the makefile.